+++ src/lib.rs
@@ -7,6 +7,9 @@ pub use configuration::Configuration;
 mod error;
 pub use error::Error;
 
+mod rate_limit;
+pub use rate_limit::{ApiResponse, RateLimit, RateLimitWindow};
+
 mod api_client;
 pub use api_client::ApiClient;